backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
transformers>=4.35.2
torch==2.9.1
pillow>=10.3.0
python-multipart>=0.0.6
pydantic>=2.10.0
python-dotenv>=1.0.0
scipy>=1.11.0
numpy>=1.24.0
